requirements.txt
argcomplete==3.5.0
asgiref==3.8.1
boto3==1.35.7
botocore==1.35.7
certifi==2024.8.30
cffi==1.17.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.7.1
dj-database-url==2.0.0
dj-rest-auth==4.0.0
Django==5.1
django-allauth==0.52.0
django-cors-headers==4.1.0
django-filter==23.2
django-social-share==2.3.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
durationpy==0.7
hjson==3.1.0
idna==3.8
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
mypy-extensions==1.0.0
mysql-connector-python==8.0.31
oauthlib==3.2.2
pillow==10.4.0
placebo==0.9.0
protobuf==3.20.1
psycopg2-binary==2.9.7
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.5
troposphere==4.8.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
zappa==0.59.0
